# .json_examples/openapi.json
{"openapi":"3.0.3","info":{"title":"R Data Core Admin API","description":"Administrative interface for the Digital Asset Management backend","contact":{"name":"Bent Brüggemann","email":"mail@bent-brueggemann.de"},"license":{"name":"MIT","url":"https://opensource.org/licenses/MIT"},"version":"0.1.0"},"servers":[{"url":"http://localhost:8888","description":"Local development server"}],"paths":{"/admin/api/v1/auth/login":{"post":{"tags":["admin-auth"],"summary":"Login endpoint for admin users","operationId":"admin_login","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AdminLoginRequest"}}},"required":true},"responses":{"200":{"description":"Login successful. Copy the token and click the Authorize button at the top to use it.","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AdminLoginResponse"}}}},"400":{"description":"Invalid request format or missing JSON body"},"401":{"description":"Invalid credentials"},"403":{"description":"Account locked or inactive"},"422":{"description":"Missing or invalid required fields"},"500":{"description":"Internal server error"}},"security":[]}},"/admin/api/v1/auth/logout":{"post":{"tags":["admin-auth"],"summary":"Logout endpoint for admin users","operationId":"admin_logout","responses":{"200":{"description":"Logout successful"},"400":{"description":"Invalid request format"},"401":{"description":"Unauthorized, invalid or missing token"},"500":{"description":"Internal server error"}},"security":[{"jwt":[]}]}},"/admin/api/v1/auth/register":{"post":{"tags":["admin-auth"],"summary":"Register a new admin user endpoint","operationId":"admin_register","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AdminRegisterRequest"}}},"required":true},"responses":{"201":{"description":"Registration successful"},"400":{"description":"Invalid request format or missing JSON body"},"403":{"description":"Insufficient permissions"},"422":{"description":"Missing or invalid required fields"},"500":{"description":"Internal server error"}},"security":[{"jwt":[]}]}},"/admin/api/v1/entity-definitions":{"get":{"tags":["admin"],"summary":"List entity definitions","operationId":"list_entity_definitions","parameters":[{"name":"limit","in":"query","description":"Maximum number of items to return","required":false,"schema":{"type":"integer","format":"int64","nullable":true}},{"name":"offset","in":"query","description":"Number of items to skip","required":false,"schema":{"type":"integer","format":"int64","nullable":true}}],"responses":{"200":{"description":"List of entity definitions","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/EntityDefinition"}}}}},"500":{"description":"Internal server error"}},"security":[{"bearerAuth":[]}]},"post":{"tags":["admin"],"summary":"Create a new entity definition","operationId":"create_entity_definition","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/EntityDefinition"}}},"required":true},"responses":{"201":{"description":"Class definition created successfully"},"500":{"description":"Internal server error"}},"security":[{"bearerAuth":[]}]}},"/admin/api/v1/entity-definitions/{uuid}":{"get":{"tags":["admin"],"summary":"Get a entity definition by UUID","operationId":"get_entity_definition","parameters":[{"name":"uuid","in":"path","description":"Class definition UUID","required":true,"schema":{"$ref":"#/components/schemas/Uuid"}}],"responses":{"200":{"description":"Class definition found","content":{"application/json":{"schema":{"$ref":"#/components/schemas/EntityDefinition"}}}},"404":{"description":"Class definition not found"},"500":{"description":"Internal server error"}},"security":[{"bearerAuth":[]}]},"put":{"tags":["admin"],"summary":"Update a entity definition","operationId":"update_entity_definition","parameters":[{"name":"uuid","in":"path","description":"Class definition UUID","required":true,"schema":{"$ref":"#/components/schemas/Uuid"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/EntityDefinition"}}},"required":true},"responses":{"200":{"description":"Class definition updated successfully"},"404":{"description":"Class definition not found"},"500":{"description":"Internal server error"}},"security":[{"bearerAuth":[]}]},"delete":{"tags":["admin"],"summary":"Delete a entity definition","operationId":"delete_entity_definition","parameters":[{"name":"uuid","in":"path","description":"Class definition UUID","required":true,"schema":{"$ref":"#/components/schemas/Uuid"}}],"responses":{"200":{"description":"Class definition deleted successfully"},"400":{"description":"Cannot delete entity definition with existing entities"},"404":{"description":"Class definition not found"},"500":{"description":"Internal server error"}},"security":[{"bearerAuth":[]}]}}},"components":{"schemas":{"AdminLoginRequest":{"type":"object","description":"Admin login request body","required":["username","password"],"properties":{"password":{"type":"string","description":"Password"},"username":{"type":"string","description":"Username or email"}}},"AdminLoginResponse":{"type":"object","description":"Admin login response body","required":["token","user_uuid","username","role","expires_at"],"properties":{"expires_at":{"type":"integer","format":"int64","description":"Token expiration (timestamp)"},"role":{"type":"string","description":"User role"},"token":{"type":"string","description":"JWT token"},"user_uuid":{"type":"string","description":"User UUID"},"username":{"type":"string","description":"Username"}}},"AdminRegisterRequest":{"type":"object","description":"Admin registration request body","required":["username","email","password","first_name","last_name"],"properties":{"email":{"type":"string","description":"Email"},"first_name":{"type":"string","description":"First name"},"last_name":{"type":"string","description":"Last name"},"password":{"type":"string","description":"Password"},"role":{"type":"string","description":"User role","nullable":true},"username":{"type":"string","description":"Username"}}},"AdminRegisterResponse":{"type":"object","description":"Admin registration response body","required":["uuid","username","message"],"properties":{"message":{"type":"string","description":"Message"},"username":{"type":"string","description":"Username"},"uuid":{"type":"string","description":"User UUID"}}},"EmptyRequest":{"type":"object","description":"Empty request body for endpoints that don't require any input"},"PaginationQuery":{"type":"object","description":"Pagination query parameters","properties":{"limit":{"type":"integer","format":"int64","nullable":true},"offset":{"type":"integer","format":"int64","nullable":true}}},"PathUuid":{"type":"object","description":"Path parameter for UUID","required":["uuid"],"properties":{"uuid":{"$ref":"#/components/schemas/Uuid"}}}},"securitySchemes":{"jwt":{"type":"http","scheme":"bearer","bearerFormat":"JWT"}}},"tags":[{"name":"admin-auth","description":"Admin authentication endpoints"},{"name":"admin","description":"Administrative endpoints"}]}